client-tracing/requirements.txt
ipfshttpclient==0.8.0a2
requests==2.27.1
web3==6.5.0
eth-utils==2.1.1
py-solc-x==1.1.1
inquirerpy==0.3.4